fix: progress bar
